JAVASCRIPT ARRAYS AND ARRAY METHODS.
learning JavaScript has been awesome. My initial commit on this space was JavaScript variables, objects and object methods and arrays.
i have been learning arrays and its method.And I must say that I haven't felt this good for awhile. A big thanks to The Odin project,  w3schools, javascript.info and a whole bunch of platforms that has provided the helpful resources in this course.